ensure git is installed for the build-n-test script
In an effort to make the build-n-test script work immediately within a
fresh studio, have the script confirm git is present or install it if
not.

`type` is a shell built-in and will error if the command given is not
present on the PATH. So build hosts that already have a git installed
will continue to work, and studio clean rooms that do not will take care
of themselves.

Signed-off-by: Robb Kidd <robb@thekidds.org>